Creation parent class for Point and EditPoint
